+++ examples/zephyr/apps/memfault_demo_app/src/main.c
@@ -4,20 +4,18 @@
 //! See License.txt for details
 
 #include <zephyr.h>
+#include <logging/log.h>
+LOG_MODULE_REGISTER(mflt, LOG_LEVEL_DBG);
 
 #include <string.h>
 
-// Product-specific Memfault dependencies
-#include "memfault/core/build_info.h"
-#include "memfault/core/math.h"
-#include "memfault/core/platform/device_info.h"
-#include "memfault/http/http_client.h"
+#include "memfault/components.h"
+
 #include "memfault/ports/zephyr/http.h"
 
 sMfltHttpClientConfig g_mflt_http_client_config = {
   .api_key = "<YOUR PROJECT KEY HERE>",
 };
-
 
 static char s_fw_version[16]="1.0.0+";
 
@@ -39,13 +37,14 @@ void memfault_platform_get_device_info(sMemfaultDeviceInfo *info) {
 
   *info = (sMemfaultDeviceInfo) {
     .device_serial = "DEMOSERIAL",
-    .software_type = "zephyr-main",
+    .software_type = "zephyr-app",
     .software_version = s_fw_version,
-    .hardware_version = "disco_l475_iot1",
+    .hardware_version = CONFIG_BOARD,
   };
 }
 
 void main(void) {
-  printk("Memfault Demo App! Board %s\n", CONFIG_BOARD);
+  LOG_INF("Memfault Demo App! Board %s\n", CONFIG_BOARD);
+  memfault_device_info_dump();
   memfault_zephyr_port_install_root_certs();
 }