+++ tests/mocks/mock_memfault_platform_debug_log.cpp
@@ -15,10 +15,6 @@
 #include "memfault/core/log.h"
 #include "memfault/core/compiler.h"
 
-// stub
-void memfault_log_save(MEMFAULT_UNUSED eMemfaultPlatformLogLevel level,
-                       MEMFAULT_UNUSED const char *fmt, ...) { }
-
 #define LOG_BUFFER_SIZE (512)
 
 void memfault_platform_log(eMemfaultPlatformLogLevel level, const char *fmt, ...) {